I made the prices of buildings doubles so that the price chantges word

+++ src/Building.java
@@ -12,21 +12,21 @@ public class Building {
         return baseRate * count;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         if (name.equals("Clicker")) {
-            return 10 * (int) Math.pow(1.15, count);
+            return 10 * Math.pow(1.15, count);
         } else if (name.equals("Grandma")) {
-            return 100 * (int) Math.pow(1.15, count);
+            return 100 * Math.pow(1.15, count);
         } else if (name.equals("Farm")) {
-            return 1000 * (int) Math.pow(1.15, count);
+            return 1000 * Math.pow(1.15, count);
         } else if (name.equals("Factory")) {
-            return 10000 * (int) Math.pow(1.15, count);
+            return 10000 * Math.pow(1.15, count);
         } else if (name.equals("Bank")) {
-            return 1000000 * (int) Math.pow(1.15, count);
+            return 1000000 * Math.pow(1.15, count);
         } else if (name.equals("Temple")) {
-            return 1000000 * (int) Math.pow(1.15, count);
+            return 1000000 * Math.pow(1.15, count);
         } else if (name.equals("Wizard-Tower")) {
-            return 1000000000 * (int) Math.pow(1.15, count);
+            return 1000000000 * Math.pow(1.15, count);
         }
         return Integer.MAX_VALUE; // Default case
     }
